Refactored code, some tests are not working.
